+++ src/loggedfs/_core/ipc.py
@@ -85,6 +85,7 @@ class _receiver_class:
 		self._t = threading.Thread(target = decoder_func, args = (self._id, self._s, self._q))
 		self._t.daemon = True
 		self._t.start()
+		self.join = self._t.join
 
 
 	def flush(self):
@@ -151,6 +152,8 @@ class receiver_manager_class:
 			self._out_r.flush()
 			self._err_r.flush()
 			self._proc_alive = self._proc.poll() is None
+		self._out_r.join()
+		self._err_r.join()
 		self._exit_func()