+++ src/loggedfs/_core/out.py
@@ -102,7 +102,7 @@ def event(format_pattern = ''):
 					ret_value = (getattr(type(e), ATTR_NAME, NAME_UNKNOWN), e.errno)
 					raise FuseOSError(e.errno)
 				else:
-					self.logger.exception(log_msg(self._log_json, ERROR_STAGE1))
+					self._logger.exception(log_msg(self._log_json, ERROR_STAGE1))
 					raise e
 			else:
 				return ret_value
@@ -115,7 +115,7 @@ def event(format_pattern = ''):
 						ret_status, ret_value
 						)
 				except Exception as e:
-					self.logger.exception(log_msg(self._log_json, ERROR_STAGE2))
+					self._logger.exception(log_msg(self._log_json, ERROR_STAGE2))
 					raise e
 
 		return wrapped
@@ -215,7 +215,7 @@ def _log_event_(
 			arg_dict[k] = self._full_path(arg_dict[k])
 	try:
 		arg_dict['buf_len'] = len(arg_dict['buf'])
-		arg_dict['buf'] = _encode_bytes_(arg_dict['buf'])
+		arg_dict['buf'] = _encode_bytes_(arg_dict['buf']) if self._log_buffers else ''
 	except KeyError:
 		pass
 
@@ -231,8 +231,8 @@ def _log_event_(
 			)):
 			log_dict['return'] = ret_value
 		elif isinstance(ret_value, bytes):
-			log_dict['return'] = _encode_bytes_(ret_value)
 			log_dict['return_len'] = len(ret_value)
+			log_dict['return'] = _encode_bytes_(ret_value) if self._log_buffers else ''
 
 	else: # FAILURE
 		log_dict.update({
@@ -242,11 +242,11 @@ def _log_event_(
 			'return_errorcode': errno.errorcode[ret_value[1]]
 			})
 
-	if not self._filter.match(log_dict):
+	if not self._log_filter.match(log_dict):
 		return
 
 	if self._log_json:
-		self.logger.info( json.dumps(log_dict, sort_keys = True)[1:-1] )
+		self._logger.info( json.dumps(log_dict, sort_keys = True)[1:-1] )
 		return
 
 	log_out = ' '.join([
@@ -258,4 +258,4 @@ def _log_event_(
 		'( %s = %d )' % ret_value
 		])
 
-	self.logger.info(log_out)
+	self._logger.info(log_out)